+++ src/Extension/InlineProgressBarCtrl.h
@@ -1,6 +1,5 @@
 #pragma once
-#include <windows.h>
-#include <commctrl.h>
+#include "stdafx.h"
 
 HWND InlineProgressBarCtrl_Create(HWND hwndParent, const long nCurrentValue, const long MaxValue, const BOOL bSmooth, const int nPane);
 BOOL InlineProgressBarCtrl_SetRange(HWND hwnd, const long nLower, const long nUpper, const long nStep);
