+++ src/Extension/stdafx.h
@@ -1,0 +1,14 @@
+#pragma once
+#ifndef _WIN32_WINNT
+#define _WIN32_WINNT 0x501
+#endif
+#ifndef NTDDI_VERSION
+#define NTDDI_VERSION NTDDI_WINXP
+#endif
+#include <windows.h>
+#include <shlobj.h>
+#include <shlwapi.h>
+#include <commctrl.h>
+#include <uxtheme.h>
+#include <stdio.h>
+#include <string.h>
