+++ src/Helpers.h
@@ -1,7 +1,4 @@
 #pragma once
-#ifdef __cplusplus
-extern "C" {
-#endif
 /******************************************************************************
 *
 *
@@ -22,20 +19,15 @@ extern "C" {
 *
 ******************************************************************************/
 
+
+
 extern HINSTANCE g_hInstance;
 extern UINT16 g_uWinVer;
 
+
 #define COUNTOF(ar) (sizeof(ar)/sizeof(ar[0]))
 #define CSTRLEN(s)  (COUNTOF(s)-1)
-#define STRING2(x) #x
-#define STRING(x) STRING2(x)
-#define CT_WARNING(M) message("\n\t** " M " => [ "__FILE__ ":" STRING(__LINE__) " ]  **\n")
-enum CSS_PROP
-{
-  css_prop_sassy = 1 << 0,
-  css_prop_less = 1 << 1,
-  css_prop_hss = 1 << 2,
-};
+
 
 extern WCHAR szIniFile[MAX_PATH];
 #define IniGetString(lpSection,lpName,lpDefault,lpReturnedStr,nSize) \
@@ -66,8 +58,8 @@ __inline BOOL IniSectionSetInt(LPWSTR lpCachedIniSection, LPCWSTR lpName, int i)
   return IniSectionSetString(lpCachedIniSection, lpName, tch);
 }
 
-extern HWND hwndEdit;
 
+extern HWND hwndEdit;
 __inline void BeginWaitCursor()
 {
   SendMessage(hwndEdit, SCI_SETCURSOR, (WPARAM)SC_CURSORWAIT, 0);
@@ -80,24 +72,28 @@ __inline void EndWaitCursor()
   SetCursorPos(pt.x, pt.y);
 }
 
+
 #define Is2k()    (g_uWinVer >= 0x0500)
 #define IsXP()    (g_uWinVer >= 0x0501)
 #define IsVista() (g_uWinVer >= 0x0600)
 #define IsW7()    (g_uWinVer >= 0x0601)
 
-#define	HL_MAX_PATH_N_CMD_LINE	MAX_PATH + 40
 
 BOOL PrivateIsAppThemed();
 HRESULT PrivateSetCurrentProcessExplicitAppUserModelID(PCWSTR);
 BOOL IsElevated();
+
+
 BOOL BitmapMergeAlpha(HBITMAP, COLORREF);
 BOOL BitmapAlphaBlend(HBITMAP, COLORREF, BYTE);
 BOOL BitmapGrayScale(HBITMAP);
 BOOL VerifyContrast(COLORREF, COLORREF);
 BOOL IsFontAvailable(LPCWSTR);
 
+
 BOOL SetWindowTitle(HWND, UINT, BOOL, UINT, LPCWSTR, int, BOOL, UINT, BOOL, LPCWSTR);
 void SetWindowTransparentMode(HWND, BOOL);
+
 
 void CenterDlgInParent(HWND);
 void GetDlgPos(HWND, LPINT, LPINT);
@@ -110,6 +106,7 @@ HDWP DeferCtlPos(HDWP, HWND, int, int, int, UINT);
 void MakeBitmapButton(HWND, int, HINSTANCE, UINT);
 void MakeColorPickButton(HWND, int, HINSTANCE, COLORREF);
 void DeleteBitmapButton(HWND, int);
+
 
 #define StatusSetSimple(hwnd,b) SendMessage(hwnd,SB_SIMPLE,(WPARAM)b,0)
 BOOL StatusSetText(HWND, UINT, LPCWSTR);
@@ -129,20 +126,24 @@ LRESULT SendWMSize(HWND);
 
 BOOL IsCmdEnabled(HWND, UINT);
 
+
 #define GetString(id,pb,cb) LoadString(g_hInstance,id,pb,cb)
 
 #define StrEnd(pStart) (pStart + lstrlen(pStart))
 
 int FormatString(LPWSTR, int, UINT, ...);
 
+
 void PathRelativeToApp(LPWSTR, LPWSTR, int, BOOL, BOOL, BOOL);
 void PathAbsoluteFromApp(LPWSTR, LPWSTR, int, BOOL);
+
 
 BOOL PathIsLnkFile(LPCWSTR);
 BOOL PathGetLnkPath(LPCWSTR, LPWSTR, int);
 BOOL PathIsLnkToDirectory(LPCWSTR, LPWSTR, int);
 BOOL PathCreateDeskLnk(LPCWSTR);
 BOOL PathCreateFavLnk(LPCWSTR, LPCWSTR, LPCWSTR);
+
 
 BOOL StrLTrim(LPWSTR, LPCWSTR);
 BOOL TrimString(LPWSTR);
@@ -153,13 +154,16 @@ void PrepareFilterStr(LPWSTR);
 void StrTab2Space(LPWSTR);
 void PathFixBackslashes(LPWSTR);
 
+
 void  ExpandEnvironmentStringsEx(LPWSTR, DWORD);
 void  PathCanonicalizeEx(LPWSTR);
 DWORD GetLongPathNameEx(LPWSTR, DWORD);
 DWORD_PTR SHGetFileInfo2(LPCWSTR, DWORD, SHFILEINFO *, UINT, UINT);
 
+
 int  FormatNumberStr(LPWSTR);
 BOOL SetDlgItemIntEx(HWND, int, UINT);
+
 
 #define MBCSToWChar(c,a,w,i) MultiByteToWideChar(c,0,a,-1,w,i)
 #define WCharToMBCS(c,w,a,i) WideCharToMultiByte(c,0,w,-1,a,i,NULL,NULL)
@@ -168,7 +172,9 @@ UINT    GetDlgItemTextA2W(UINT, HWND, int, LPSTR, int);
 UINT    SetDlgItemTextA2W(UINT, HWND, int, LPSTR);
 LRESULT ComboBox_AddStringA2W(UINT, HWND, LPCSTR);
 
+
 UINT CodePageFromCharSet(UINT);
+
 
 //==== MRU Functions ==========================================================
 #define MRU_MAXITEMS 24
@@ -181,6 +187,7 @@ typedef struct _mrulist
   int   iFlags;
   int   iSize;
   LPWSTR pszItems[MRU_MAXITEMS];
+
 } MRULIST, *PMRULIST, *LPMRULIST;
 
 LPMRULIST MRU_Create(LPCWSTR, int, int);
@@ -194,6 +201,7 @@ int       MRU_Enum(LPMRULIST, int, LPWSTR, int);
 BOOL      MRU_Load(LPMRULIST);
 BOOL      MRU_Save(LPMRULIST);
 BOOL      MRU_MergeSave(LPMRULIST, BOOL, BOOL, BOOL);
+
 
 //==== Themed Dialogs =========================================================
 #ifndef DLGTEMPLATEEX
@@ -221,81 +229,16 @@ DLGTEMPLATE *LoadThemedDialogTemplate(LPCTSTR, HINSTANCE);
 INT_PTR ThemedDialogBoxParam(HINSTANCE, LPCTSTR, HWND, DLGPROC, LPARAM);
 HWND    CreateThemedDialogParam(HINSTANCE, LPCTSTR, HWND, DLGPROC, LPARAM);
 
+
 //==== UnSlash Functions ======================================================
 void TransformBackslashes(char *, BOOL, UINT);
+
 
 //==== MinimizeToTray Functions - see comments in Helpers.c ===================
 BOOL GetDoAnimateMinimize(VOID);
 VOID MinimizeWndToTray(HWND hWnd);
 VOID RestoreWndFromTray(HWND hWnd);
 
-//============== haccel work
-VOID	HL_Init(HWND hwnd);
-VOID	HL_LoadINI();
-VOID	HL_SaveINI();
-VOID	HL_Release();
-VOID	HL_Trace(const char *fmt, ...);
-VOID	HL_WTrace(const char *fmt, LPCWSTR word);
-VOID	HL_WTrace2(const char *fmt, LPCWSTR word1, LPCWSTR word2);
-BOOL	HL_Is_Empty(LPCWSTR txt);
-BOOL	HL_Get_goto_number(LPTSTR txt, int *out, BOOL hex);
-VOID	HL_Set_wheel_scroll(BOOL on);
-VOID	HL_Reload_Settings();
-VOID	HL_Get_last_dir(LPTSTR out);
-void	HL_inplace_rev(WCHAR * s);
-BOOL	HL_Explorer_cxt_menu(LPCWSTR path, void * parentWindow);
-UINT_PTR CALLBACK HL_OFN__hook_proc(
-  HWND hdlg,
-  UINT uiMsg,
-  WPARAM wParam,
-  LPARAM lParam
-  );
 
-VOID	HL_Grep(VOID* lpf, BOOL grep);
 
-void*	HL_Alloc(size_t);
-void*	HL_Realloc(void*, size_t);
-void	HL_Free(void*);
-int		HL_Compare_files(LPCWSTR sz1, LPCWSTR sz2);
-BOOL	hl_iswordchar(WCHAR ch);
-BOOL	hl_isspace(WCHAR ch);
-
-#define HL_INI_SECTION L"Notepad2e"
-#define INI_SETTING_HIGHLIGHT_LINE_IF_WINDOW_INACTIVE L"HighlightLineIfWindowInactive"
-#define INI_SETTING_SCROLL_Y_CARET_POLICY L"ScrollYCaretPolicy"
-#define INI_SETTING_FIND_WORD_MATCH_CASE  L"FindWordMatchCase"
-#define INI_SETTING_FIND_WRAP_AROUND  L"FindWordWrapAround"
-#define INI_SETTING_MOVE_CARET_ON_RIGHT_CLICK L"MoveCaretOnRightClick"
-#define INI_SETTING_MATH_EVAL L"MathEval"
-#define INI_SETTING_LANGUAGE_INDICATOR L"TitleLanguage"
-#define INI_SETTING_WORD_NAVIGATION_MODE  L"WordNavigationMode"
-
-#define HWM_RELOAD_SETTINGS	(WM_USER + 0xee)
-#ifdef _DEBUG
-#define __FILE_LOC (1+strrchr(__FILE__,'\\')/*?strrchr(__FILE__,'\\')+1:__FILE__*/)
-#define HL_TRACE(FMT,...)	HL_Trace ( "[%s: %d] - "#FMT , __FILE_LOC , __LINE__ , __VA_ARGS__ );
-#define HL_TRACE_S(OBJ)		HL_Trace ( "[%s: %d] [%s]=%s " , __FILE_LOC , __LINE__ , #OBJ , OBJ );
-#define HL_TRACE_I(OBJ)		HL_Trace ( "[%s: %d] [%s]=%d (0x%04xd) " , __FILE_LOC , __LINE__ , #OBJ , OBJ , OBJ );
-#define HL_TRACE_TR(OBJ)		HL_Trace ( "[%s: %d] [%s]= TEXTRANGE %d:%d(%s) " , __FILE_LOC , __LINE__ , #OBJ , OBJ.chrg.cpMin ,OBJ.chrg.cpMax ,OBJ.lpstrText );
-#else
-#define HL_TRACE(FMT,...)	 (void)(FMT);
-#define HL_TRACE_S(OBJ)		(void)(OBJ);
-#define HL_TRACE_I(OBJ)		(void)(OBJ);
-#define HL_TRACE_TR(OBJ)		(void)(OBJ);
-#endif
-
-BOOL SetClipboardText(const HWND hwnd, const wchar_t* text);
-BOOL ASCIItoUCS2(const char* lpSrc, wchar_t* lpDest, const int maxDest, const UINT nCodePage);
-
-typedef enum
-{
-  ELI_HIDE = 0,
-  ELI_SHOW,
-  ELI_SHOW_NON_US
-} ELanguageIndicatorMode;
-
-void UpdateWindowTitle(HWND hwnd);
-#ifdef __cplusplus
-}//end extern "C"
-#endif
 ///   End of Helpers.h   \\\
